6.1.0 + Registry tweak to adjust date format
Added code to change date format via registry command, and to create a batch file to restore it.
Added function to run restoration batch and then delete it.
Added function calls at all script exit points that I could identify.
Replaced all instances of %CUR_DATE% with %DATE%.
Removed CUR_DATE definition.

Left in DTS definition for use in LOGPATH/LOGFILE, but modified it to exclude decimal seconds and the trailing info. If you'd rather keep the decimal seconds, that's easy enough, but I really recommend ditching the trailing info.

Batch file allows for restoring format manually if Tron is interrupted. To make manual restoration easier for inexperienced users, we may want to add a shortcut on the desktop, or add explanatory text in the intro screens. At a minimum, something something should be added to the Instructions file.